moving sprites properly, and rendering background

+++ aya-console/src/renderer/mod.rs
@@ -6,7 +6,7 @@ use error::Result;
 pub use raylib::RaylibRenderer;
 
 pub trait Renderer {
-    fn start(name: &str, fps: f64, scale: u16) -> Self;
+    fn start(name: &str, fps: f32, scale: u16) -> Self;
     fn should_close(&self) -> bool;
     fn should_draw(&self) -> bool;
     fn draw_frame(&mut self, memory: &mut impl Addressable) -> Result<()>;
